requirements.txt
-e ./
filelock
flake8>=3.8.4
pep8
pyflakes
black>=20.8b1
pytest-cov>=2.6.1
coverage>=5.1
pytest
cython
sympy
versioneer
jax
jaxlib
numba>=0.55
numba-scipy>=0.3.0
diff-cover
pre-commit
isort
packaging
typing_extensions
pytest-benchmark
